+++ fiddles/nativeScript/fiddle-0006-Plugins/chat/platforms/android/src/main/assets/app/tns_modules/nativescript-telerik-ui/sidedrawer/drawerpage/index.d.ts
@@ -1,0 +1,12 @@
+import page = require("ui/page");
+import drawer = require('../../sidedrawer');
+
+/**
+ * DrawerPage allows developers to place RadSideDrawer over the navigation bar/action bar.
+ */
+export class DrawerPage extends page.Page {
+    /**
+     * The side drawer of DrawerPage.
+     */
+    sideDrawer: drawer.RadSideDrawer;
+}